Create TSP (Traveling Salesman Problem)
Description of the problem: 
A tourist wants to visit n cities 1,2,..,n. A journey is a way of traveling starting from a certain city, going through all the remaining cities, each city exactly once, then returning to the starting city. Knowing cost[i][j] is the cost of traveling from city i to city j (i, j = 1, 2,..., n), Find the journey with the smallest total cost.
Input is the number of cities and the cost matrix (Ex: cost[2][5] = 1 <=> cost to travel from city 1 to city 5 is 1)