Create Knapsack Problem
Description:  You have a set of items (knapsacks) with different values and weights. Each item has an unlimited supply, meaning you can take as many units of it as you want. However, you have a maximum weight limit that you cannot exceed. The goal is to select the combination of items that maximizes the total value without exceeding the weight limit. The result will be the number of each item we should select.